Correccion y optimizacion de endpoint

+++ src/main/resources/db/migration/V1__init.sql
@@ -1,12 +1,12 @@
 CREATE TABLE IF NOT EXISTS categorias (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nombre VARCHAR(100) NOT NULL,
     descripcion TEXT
 );
 
 CREATE TABLE IF NOT EXISTS platos (
-    id SERIAL PRIMARY KEY,
-    categoria_id INTEGER NOT NULL REFERENCES categorias(id),
+    id BIGSERIAL PRIMARY KEY,
+    categoria_id BIGINT NOT NULL REFERENCES categorias(id),
     nombre VARCHAR(200) NOT NULL,
     descripcion TEXT,
     precio DECIMAL(10, 2) NOT NULL,
@@ -14,7 +14,7 @@ CREATE TABLE IF NOT EXISTS platos (
 );
 
 CREATE TABLE IF NOT EXISTS mesas (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     numero INTEGER NOT NULL UNIQUE,
     capacidad INTEGER NOT NULL,
     ubicacion VARCHAR(100),
@@ -22,14 +22,14 @@ CREATE TABLE IF NOT EXISTS mesas (
 );
 
 CREATE TABLE IF NOT EXISTS clientes (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nombre VARCHAR(200) NOT NULL,
     telefono VARCHAR(20),
     email VARCHAR(200)
 );
 
 CREATE TABLE IF NOT EXISTS usuarios (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nombre VARCHAR(200) NOT NULL,
     rol VARCHAR(50) NOT NULL,
     usuario VARCHAR(100) NOT NULL UNIQUE,
@@ -37,26 +37,26 @@ CREATE TABLE IF NOT EXISTS usuarios (
 );
 
 CREATE TABLE IF NOT EXISTS pedidos (
-    id SERIAL PRIMARY KEY,
-    mesa_id INTEGER NOT NULL REFERENCES mesas(id),
-    cliente_id INTEGER REFERENCES clientes(id),
-    usuario_id INTEGER REFERENCES usuarios(id),
+    id BIGSERIAL PRIMARY KEY,
+    mesa_id BIGINT NOT NULL REFERENCES mesas(id),
+    cliente_id BIGINT REFERENCES clientes(id),
+    usuario_id BIGINT REFERENCES usuarios(id),
     fecha TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     estado VARCHAR(20) DEFAULT 'activo',
     total DECIMAL(10, 2) DEFAULT 0
 );
 
 CREATE TABLE IF NOT EXISTS detalle_pedido (
-    id SERIAL PRIMARY KEY,
-    pedido_id INTEGER NOT NULL REFERENCES pedidos(id),
-    plato_id INTEGER NOT NULL REFERENCES platos(id),
+    id BIGSERIAL PRIMARY KEY,
+    pedido_id BIGINT NOT NULL REFERENCES pedidos(id),
+    plato_id BIGINT NOT NULL REFERENCES platos(id),
     cantidad INTEGER NOT NULL,
     subtotal DECIMAL(10, 2) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS facturas (
-    id SERIAL PRIMARY KEY,
-    pedido_id INTEGER NOT NULL UNIQUE REFERENCES pedidos(id),
+    id BIGSERIAL PRIMARY KEY,
+    pedido_id BIGINT NOT NULL UNIQUE REFERENCES pedidos(id),
     monto DECIMAL(10, 2) NOT NULL,
     metodo_pago VARCHAR(50),
     fecha TIMESTAMP DEFAULT CURRENT_TIMESTAMP
